+++ test/services/dumper/dumper-mongo.test.js
@@ -17,9 +17,10 @@ const context = require('../../../context');
 const initContext = require('../../../context/init');
 
 initContext(context);
+const injectedContext = context.inject();
 
 function getDumper() {
-  return new Dumper(context.inject());
+  return new Dumper(injectedContext);
 }
 
 const CONFIG = {
@@ -42,13 +43,16 @@ async function getGeneratedFileFromPersonModel(model) {
   return fs.readFileSync('./test-output/mongo/models/persons.js', 'utf8');
 }
 
+const TEST_OUTPUT_MODEL_FILMS_PATH = './test-output/mongo/models/films.js';
+const TEST_EXPECTED_MODEL_FILMS_PATH = './test-expected/mongo/dumper-output/simple.expected.js';
+
 describe('services > dumper > MongoDB', () => {
   it('should generate a simple model file', async () => {
     expect.assertions(1);
     const dumper = getDumper();
     await dumper.dump(simpleModel, CONFIG);
-    const generatedFile = fs.readFileSync('./test-output/mongo/models/films.js', 'utf8');
-    const expectedFile = fs.readFileSync('./test-expected/mongo/dumper-output/simple.expected.js', 'utf-8');
+    const generatedFile = fs.readFileSync(TEST_OUTPUT_MODEL_FILMS_PATH, 'utf8');
+    const expectedFile = fs.readFileSync(TEST_EXPECTED_MODEL_FILMS_PATH, 'utf8');
 
     expect(generatedFile).toStrictEqual(expectedFile);
     cleanOutput();
@@ -58,18 +62,43 @@ describe('services > dumper > MongoDB', () => {
     expect.assertions(1);
     const dumper = getDumper();
     await dumper.dump(hasManyModel, CONFIG);
-    const generatedFile = fs.readFileSync('./test-output/mongo/models/films.js', 'utf8');
+    const generatedFile = fs.readFileSync(TEST_OUTPUT_MODEL_FILMS_PATH, 'utf8');
     const expectedFile = fs.readFileSync('./test-expected/mongo/dumper-output/hasmany.expected.js', 'utf-8');
 
     expect(generatedFile).toStrictEqual(expectedFile);
-    // cleanOutput();
+    cleanOutput();
   });
 
   describe('handling /models/index.js file', () => {
     it('should not force type casting', async () => {
       expect.assertions(1);
+      const dumper = await getDumper();
+      await dumper.dump(simpleModel, CONFIG);
       const indexGeneratedFile = fs.readFileSync('./test-output/mongo/models/index.js', 'utf-8');
+
       expect(indexGeneratedFile).toStrictEqual(expect.not.stringMatching('databaseOptions.dialectOptions.typeCast'));
+      cleanOutput();
+    });
+
+    it('should generate a model/index.js file', async () => {
+      expect.assertions(1);
+      const dumper = await getDumper();
+      await dumper.dump(simpleModel, CONFIG);
+      const indexGeneratedFile = fs.readFileSync('./test-output/mongo/models/index.js', 'utf-8');
+      const expectedFile = fs.readFileSync('./test-expected/mongo/dumper-output/index.expected.js', 'utf-8');
+
+      expect(indexGeneratedFile).toStrictEqual(expectedFile);
+      cleanOutput();
+    });
+
+    it('should generate a config/databases.js file', async () => {
+      expect.assertions(1);
+      const dumper = await getDumper();
+      await dumper.dump(simpleModel, CONFIG);
+      const indexGeneratedFile = fs.readFileSync('./test-output/mongo/config/databases.js', 'utf-8');
+      const expectedFile = fs.readFileSync('./test-expected/mongo/dumper-output/databases.config.expected.js', 'utf-8');
+
+      expect(indexGeneratedFile).toStrictEqual(expectedFile);
       cleanOutput();
     });
   });
@@ -153,5 +182,24 @@ describe('services > dumper > MongoDB', () => {
 
     expect(generatedFile).toStrictEqual(expectedFile);
     cleanOutput();
+  });
+
+  describe('on re-dump', () => {
+    it('should recreate files that have been deleted', async () => {
+      expect.assertions(1);
+
+      // Setup test by dumping once to ensure all files exists, then remove a file
+      const dumper = getDumper();
+      await dumper.dump(simpleModel, CONFIG);
+      fs.unlinkSync(TEST_OUTPUT_MODEL_FILMS_PATH);
+
+      await dumper.dump(simpleModel, { ...CONFIG, isUpdate: true });
+      const generatedFile = fs.readFileSync(TEST_OUTPUT_MODEL_FILMS_PATH, 'utf8');
+      const expectedFile = fs.readFileSync(TEST_EXPECTED_MODEL_FILMS_PATH, 'utf8');
+
+      // Then we ensure that the file that were removed exists after a redump
+      expect(generatedFile).toStrictEqual(expectedFile);
+      cleanOutput();
+    });
   });
 });