+++ test/services/dumper/dumper-sequelize.test.js
@@ -39,12 +39,14 @@ function cleanOutput() {
   rimraf.sync('./test-output/sequelize');
 }
 
+const TEST_OUTPUT_MODEL_CUSTOMERS_PATH = './test-output/sequelize/models/customers.js';
+
 describe('services > dumper > sequelize', () => {
   it('should generate a simple model file', async () => {
     expect.assertions(1);
     const dumper = getDumper();
     await dumper.dump(simpleModel, CONFIG);
-    const generatedFile = fs.readFileSync('./test-output/sequelize/models/customers.js', 'utf8');
+    const generatedFile = fs.readFileSync(TEST_OUTPUT_MODEL_CUSTOMERS_PATH, 'utf8');
     const expectedFile = fs.readFileSync('./test-expected/sequelize/dumper-output/customers.expected.js', 'utf-8');
 
     expect(generatedFile).toStrictEqual(expectedFile);
@@ -157,6 +159,17 @@ describe('services > dumper > sequelize', () => {
     cleanOutput();
   });
 
+  it('should generate a config/databases.js file', async () => {
+    expect.assertions(1);
+    const dumper = await getDumper();
+    await dumper.dump(simpleModel, CONFIG);
+    const indexGeneratedFile = fs.readFileSync('./test-output/sequelize/config/databases.js', 'utf-8');
+    const expectedFile = fs.readFileSync('./test-expected/sequelize/dumper-output/databases.config.expected.js', 'utf-8');
+
+    expect(indexGeneratedFile).toStrictEqual(expectedFile);
+    cleanOutput();
+  });
+
   describe('when generating the env file on various OS', () => {
     it('should generate the env file on Linux', async () => {
       expect.assertions(1);
@@ -206,5 +219,24 @@ describe('services > dumper > sequelize', () => {
 
     expect(generatedFile).toStrictEqual(expectedFile);
     cleanOutput();
+  });
+
+  describe('on re-dump', () => {
+    it('should recreate files that have been deleted', async () => {
+      expect.assertions(1);
+
+      // Setup test by dumping once to ensure all files exists, then remove a file
+      const dumper = getDumper();
+      await dumper.dump(simpleModel, CONFIG);
+      fs.unlinkSync(TEST_OUTPUT_MODEL_CUSTOMERS_PATH);
+
+      await dumper.dump(simpleModel, { ...CONFIG, isUpdate: true });
+      const generatedFile = fs.readFileSync(TEST_OUTPUT_MODEL_CUSTOMERS_PATH, 'utf8');
+      const expectedFile = fs.readFileSync('./test-expected/sequelize/dumper-output/customers.expected.js', 'utf-8');
+
+      // Then we ensure that the file that were removed exists after a redump
+      expect(generatedFile).toStrictEqual(expectedFile);
+      cleanOutput();
+    });
   });
 });