+++ utils/errors/database/empty-database-error.js
@@ -1,0 +1,5 @@
+const LumberError = require('../../lumber-error');
+
+class EmptyDatabase extends LumberError {}
+
+module.exports = EmptyDatabase;