+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# [4.0.0-beta.5](https://github.com/ForestAdmin/lumber/compare/v4.0.0-beta.4...v4.0.0-beta.5) (2021-02-22)
+
+
+### Bug Fixes
+
+* remove merge error ([aec7593](https://github.com/ForestAdmin/lumber/commit/aec75934045b8b0e9d608967a2f4ebea309f62aa))
+* **view-list:** fix bad records count in view list if the record details access is forbidden ([#539](https://github.com/ForestAdmin/lumber/issues/539)) ([1f6f6c7](https://github.com/ForestAdmin/lumber/commit/1f6f6c7b491e33c7dd52d886a0ea7303d53b0bb6))
+* fix tests for handle foreign key toward non primary key ([#537](https://github.com/ForestAdmin/lumber/issues/537)) ([0c2ee15](https://github.com/ForestAdmin/lumber/commit/0c2ee15bc2200de352b1af58eb5624c990c961c0))
+* handle foreign key toward non primary key ([#525](https://github.com/ForestAdmin/lumber/issues/525)) ([e765af6](https://github.com/ForestAdmin/lumber/commit/e765af6e6943b4e05d800d08cb2560afd3d28ab0))
+* incorrect FOREST_URL generated in docker-compose.yml ([#533](https://github.com/ForestAdmin/lumber/issues/533)) ([9d3bf95](https://github.com/ForestAdmin/lumber/commit/9d3bf951ae743082e8ad10525734b37af1643666))
+* **models:** id primary keys on join tables are now considered and correctly added on generated models ([#524](https://github.com/ForestAdmin/lumber/issues/524)) ([50b25f8](https://github.com/ForestAdmin/lumber/commit/50b25f82d41d27c0b22ff2914720e470c24ad26b))
+
 ## [3.10.9](https://github.com/ForestAdmin/lumber/compare/v3.10.8...v3.10.9) (2021-02-10)
 
 
@@ -13,6 +25,13 @@
 * fix tests for handle foreign key toward non primary key ([#537](https://github.com/ForestAdmin/lumber/issues/537)) ([0c2ee15](https://github.com/ForestAdmin/lumber/commit/0c2ee15bc2200de352b1af58eb5624c990c961c0))
 * handle foreign key toward non primary key ([#525](https://github.com/ForestAdmin/lumber/issues/525)) ([e765af6](https://github.com/ForestAdmin/lumber/commit/e765af6e6943b4e05d800d08cb2560afd3d28ab0))
 
+# [4.0.0-beta.4](https://github.com/ForestAdmin/lumber/compare/v4.0.0-beta.3...v4.0.0-beta.4) (2021-01-12)
+
+
+### Features
+
+* **update:** add lumber update command ([#527](https://github.com/ForestAdmin/lumber/issues/527)) ([d8f1eaf](https://github.com/ForestAdmin/lumber/commit/d8f1eafc952a74f6483bdc05661dbb2bcde59e3f))
+
 ## [3.10.7](https://github.com/ForestAdmin/lumber/compare/v3.10.6...v3.10.7) (2021-01-07)
 
 
@@ -27,12 +46,38 @@
 
 * **models:** id primary keys on join tables are now considered and correctly added on generated models ([#524](https://github.com/ForestAdmin/lumber/issues/524)) ([50b25f8](https://github.com/ForestAdmin/lumber/commit/50b25f82d41d27c0b22ff2914720e470c24ad26b))
 
+# [4.0.0-beta.3](https://github.com/ForestAdmin/lumber/compare/v4.0.0-beta.2...v4.0.0-beta.3) (2021-01-06)
+
+
+### Features
+
+* ease multi-database setup by managing mongoose connection like sequelize ([#490](https://github.com/ForestAdmin/lumber/issues/490)) ([cf508cb](https://github.com/ForestAdmin/lumber/commit/cf508cb1f87a465c483086cec467494eb06e3df1))
+
+# [4.0.0-beta.2](https://github.com/ForestAdmin/lumber/compare/v4.0.0-beta.1...v4.0.0-beta.2) (2021-01-05)
+
+
+### Bug Fixes
+
+* **sequelize:** fix error when working with column table including parenthesis ([#520](https://github.com/ForestAdmin/lumber/issues/520)) ([79c041b](https://github.com/ForestAdmin/lumber/commit/79c041b8ca6a78db198e71a4b0ec1dbba224acbd))
+
 ## [3.10.5](https://github.com/ForestAdmin/lumber/compare/v3.10.4...v3.10.5) (2020-12-21)
 
 
 ### Bug Fixes
 
 * **sequelize:** fix error when working with column table including parenthesis ([#520](https://github.com/ForestAdmin/lumber/issues/520)) ([79c041b](https://github.com/ForestAdmin/lumber/commit/79c041b8ca6a78db198e71a4b0ec1dbba224acbd))
+
+# [4.0.0-beta.1](https://github.com/ForestAdmin/lumber/compare/v3.10.4...v4.0.0-beta.1) (2020-12-21)
+
+
+### Features
+
+* authenticate without entering a password ([#513](https://github.com/ForestAdmin/lumber/issues/513)) ([fe8d8c3](https://github.com/ForestAdmin/lumber/commit/fe8d8c3b0e1a45430a4b0a86a2f7e2b1a2f5d57a))
+
+
+### BREAKING CHANGES
+
+* by default, the authentication process will redirect to a web page to validate a code.
 
 ## [3.10.4](https://github.com/ForestAdmin/lumber/compare/v3.10.3...v3.10.4) (2020-12-01)